refactor tracking modes
