+++ src/tracker/fast_track.rs
@@ -1,11 +1,10 @@
+use crate::tracker::common::set_filename_and_serialize;
 use crate::RpcConnection;
 use crate::rpc::format::{hex_to_decimal, decimal_to_hex};
 use crate::tracker::types::*;
-use crate::tracker::time::get_latest_unix_timestamp;
 
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
-use std::fs;
 
 use ctrlc;
 use ethers::types::U256;
@@ -78,30 +77,7 @@ pub async fn fast_track_state(
 		current_block += interval;
 	}
 	
-	// Set the filename to `address{contract_address}-slot-{storage_slot}-timestamp-{unix_timestamp} if its the default one
-	// We also check if we should serialize it as csv
-	let json;
-	let filename = match filename.as_str() {
-		"" => {
-			let timestamp = get_latest_unix_timestamp();
-			println!("No filename specified, using default and formatting as JSON");
-			json = storage.serialize_json()?;
-			format!("address-{}-slot-{}-timestamp-{}.json", contract_address, storage_slot, timestamp)
-		},
-		filename if filename.contains(".csv") => {
-			println!("Formatting as CSV...");
-			json = storage.serialize_csv();
-			filename.to_string()
-		},
-		_ => {
-			json = storage.serialize_json()?;
-			filename
-		},
-	};
-
-	let path = format!("{}/{}", path, filename);
-	println!("\nWriting to file: {}", path);
-	fs::write(path, json)?;
+	set_filename_and_serialize(path, filename, storage, contract_address, storage_slot)?;
 
 	Ok(())
 }
